Refactor to collected validation (#88)
* Renaming fixing spelling mistake

* Remove Business validation

* Renamed input-validation to validation

* Added new validator CICD

* Removed persist tf and moved conn string to validator tf
