+++ .github/workflows/validator-cd.yml
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-name: Business Validator CD
+name: Validator CD
 
 on:
   push:
@@ -19,23 +19,23 @@ on:
       - main
     paths:
       - source/GreenEnergyHub.Charges/**
-      - .github/workflows/business-validator-cd.yml
+      - .github/workflows/validator-cd.yml
       - .github/actions/dotnet-build-and-test/**
       - .github/actions/azure-core-tools-install/**
   workflow_dispatch: {}
 
 env:
-  CSPROJ_FILE_PATH: 'source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.BusinessValidator/GreenEnergyHub.Charges.BusinessValidator.csproj'
+  CSPROJ_FILE_PATH: 'source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Validator/GreenEnergyHub.Charges.Validator.csproj'
   SOLUTION_FILE_PATH: 'source/GreenEnergyHub.Charges/GreenEnergyHub.Charges.sln'
   DOTNET_VERSION: '3.1'
   ORGANISATION_NAME: 'endk'
-  AZURE_FUNCTIONAPP_NAME: 'business-validator'
+  AZURE_FUNCTIONAPP_NAME: 'validator'
   PROJECT_NAME: charges
   REPOSITORY_NAME: Energinet-DataHub/geh-charges
 
 jobs:
-  build_business_validator:
-    name: Build Business Validator
+  build_validator:
+    name: Build Validator
     runs-on: ubuntu-latest
     steps:
       - name: Checkout repository
@@ -62,9 +62,9 @@ jobs:
           name: drop
           path: '${{ github.workspace }}/output'
     
-  deploy_business_validator:
+  deploy_validator:
     name: Deploy function to ${{ matrix.environment.name }} - ${{ matrix.environment.long }}
-    needs: build_business_validator
+    needs: build_validator
     runs-on: ubuntu-latest
     strategy:
       matrix:
