+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Domain/ChangeOfCharges/Fee/FeeCreateValidationFailed.cs
@@ -14,7 +14,7 @@
 
 namespace GreenEnergyHub.Charges.Domain.ChangeOfCharges.Fee
 {
-    public class FeeCreateInputValidationFailed : FeeBase
+    public class FeeCreateValidationFailed : FeeBase
     {
     }
 }
