+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Domain/ChangeOfCharges/Tariff/TariffCreateValidationSucceeded.cs
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace GreenEnergyHub.Charges.Domain.ChangeOfCharges.Fee
+namespace GreenEnergyHub.Charges.Domain.ChangeOfCharges.Tariff
 {
-    public class FeeCreateInputValidationSucceded : FeeBase
+    public class TariffCreateValidationSucceeded : TariffBase
     {
     }
 }
