--- source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.InputValidator/Properties/serviceDependencies.local.json
@@ -1,11 +1,0 @@
-{
-  "dependencies": {
-    "appInsights1": {
-      "type": "appInsights.sdk"
-    },
-    "storage1": {
-      "type": "storage.emulator",
-      "connectionId": "AzureWebJobsStorage"
-    }
-  }
-}
